Complete version: color sets, reset, 7 fractals, cycling, zoom ...; need to add multiwindow, shifting, legend and maybe multithreads

+++ src/fractals/buffalo.c
@@ -6,7 +6,7 @@
 /*   By: eaptekar <eaptekar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/25 16:41:45 by eaptekar          #+#    #+#             */
-/*   Updated: 2018/08/26 19:12:35 by eaptekar         ###   ########.fr       */
+/*   Updated: 2018/08/29 11:45:45 by eaptekar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ void			init_buffalo(t_fractol *f)
 	f->frac = 7;
 }
 
-static int		calcul_iter_buffalo(t_fractol *f, int x, int y)
+static double	calcul_iter_buffalo(t_fractol *f, int x, int y)
 {
 	t_calcul	mb;
 	int			n;
